requirements.txt
streamlit
pdfplumber
pymupdf
pytesseract
reportlab
pillow
python-dotenv
pypdf
